% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{fisher}
\usage{
fisher(p, adjust, R)
}
\arguments{
\item{p}{vector of pvalues}

\item{adjust}{if correlation has to be taken into account}

\item{R}{correlation matrix}
}
\description{
fisher
}
