% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianize.R
\name{gaussianize}
\alias{gaussianize}
\title{gaussianize null variable importances}
\usage{
gaussianize(x, a)
}
\arguments{
\item{x}{distr}

\item{a}{value to interpolate/extrapolate}
}
\description{
gaussianize null variable importances
}
