% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_effective.R
\name{m_effective}
\alias{m_effective}
\title{meff. written by Ozan Cinar and Wolfgang Viechtbauer.}
\usage{
m_effective(R)
}
\arguments{
\item{R}{R}
}
\description{
meff. written by Ozan Cinar and Wolfgang Viechtbauer.
}
