% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tippett.R
\name{tippett}
\alias{tippett}
\title{tippett.}
\usage{
tippett(p, adjust, R)
}
\arguments{
\item{p}{vector of pvalues}

\item{adjust}{if correlation has to be taken into account}

\item{R}{correlation matrix}
}
\description{
tippett.
}
